HTTP DELETE now sets product inventory to 0 ...
HTTP DELETE request now sets to 0 the quantity of the
of the inventory item related to the product.

Prior to this commit, an HTTP DELETE request would try
to delete the product instance but, such action would
raise an error due to the protected relationship between
the product instance and its associated inventory item
instance.

This commit fixes the issue with an appropiate behavior.
Product instances cannot be deleted because it would break
several entity relationships.