Prevent updates in InventoryItem.updated_on for PATCH request ...
Override the patch method on InventoryItemSingleView so no updates
in InventoryItem.updated_on are performed unless an actual update
on the InventoryItem instance occurs.

Previously, the inventoryItem.updated_on attribute was being updated
with PATCH requests that didn't even had a payload.

The changes provided in this commit verify that the PATCH requests
contain a valid non-empty payload and if so, the update is then performed.