Serialize correct URL for cart items
The CartSerializer was previously serializing the
URL for cart items using the cart items endpoint
for admins. This commit changes the serializer
to use the cart items endpoint for customers

The previous behavior was incorrect, as customers
are not able to access the admin cart items endpoint.
This commit fixes the issue and ensures that customers
are able to access to the serialized cart items endpoint.