Products are now soft-deleted
Products cannnot be deleted because doing so would break
very important relationships which would break the system.

Prviosly, on DELETE request, products quantity in stock was
being set to zero (0) but that was not the best solution...

This commit introduces a better apprach which is to
asocciated products with a new model called DeletedProducts
on DELETE requests.